Feat: 모듈화 추가

+++ controller/CartController.js
@@ -1,11 +1,26 @@
+const ensureAuthorization = require('../auth');
+const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
 
+
 const addToCart =  (req, res) => {
-    const {book_id, quantity, user_id} = req.body;
+    const {book_id, quantity} = req.body;
+
+    let authorization = ensureAuthorization(req, res);
+
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
 
     let sql = `INSERT INTO cartItems (book_id, quantity, user_id) VALUES (?, ?, ?);`;
-    let values = [book_id, quantity, user_id]
+    let values = [book_id, quantity, authorization.id]
     conn.query(sql, values, (err, results) => {
         if (err) {
             console.log(err);
@@ -14,43 +29,72 @@ const addToCart =  (req, res) => {
     
         return res.status(StatusCodes.OK).json(results);
         })
+    }
 };
 
 const getCartItems =(req, res) => {
-    const {user_id, selected} = req.body;
+    const {selected} = req.body;
 
-    let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
+    let authorization = ensureAuthorization(req, res);
+
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
                 FROM cartItems LEFT JOIN books 
                 ON cartItems.book_id = books.id
-                WHERE user_id = ? AND cartItems.id IN (?)`;
+                WHERE user_id = ?`;
 
-    let values = [user_id, selected]
+        let values = [authorization.id,]
 
-    conn.query(sql, values, 
-        (err, results) => {
+        if (selected) {
+            sql += ` AND cartItems.id IN (?)`;
+            values.push(selected);
+        } 
+
+        conn.query(sql, values, 
+            (err, results) => {
+                if (err) {
+                    console.log(err);
+                    return res.status(StatusCodes.BAD_REQUEST).end();
+                }
+    
+                return res.status(StatusCodes.OK).json(results);
+        })
+    }
+};
+
+const removeCartItem = (req, res) => {
+    let authorization = ensureAuthorization(req, res);
+
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const cartItemId = req.params.id;
+
+        let sql = `DELETE FROM cartItems WHERE id = ?`;
+        conn.query(sql, cartItemId, (err, results) => {
             if (err) {
                 console.log(err);
                 return res.status(StatusCodes.BAD_REQUEST).end();
             }
     
             return res.status(StatusCodes.OK).json(results);
-        })
-};
-
-const removeCartItem = (req, res) => {
-    const {id} = req.params;
-
-    let sql = `DELETE FROM cartItems WHERE id = ?`;
-    conn.query(sql, id, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
+            })
         }
-    
-        return res.status(StatusCodes.OK).json(results);
-        })
 };
-
 
 module.exports = {
     addToCart,
